- migrate feature - item placements
