--- features/itemPlacements/src/main/AndroidManifest.xml
@@ -1,6 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <uses-permission android:name="android.permission.VIBRATE" />
-
-
-</manifest>
